Allow testing of prerequisites.
The modules here allow for testing of undeclared dependencies. This is
something which we have struggled a bit with regarding Bagger.  The
onemoredata/bagger#111 PR can use these modules.

These modules are technically optional but I think we should standardize
on their use for authors at least to make sure we don;t have more
problems of missing dependencies.